ISSU-33 Add conditional rendering to Apply button
